moving script running commands to separate file so can be called from cli
